Updated project to build successfully
I added the Twitter and Account Frameworks to the project so that the
code would link. I also had to add the UMATwitterAPI.m file to the list
of sources to compile.

I cleaned up some code in UMATwitterAPI.m to import the Twitter and
Account headers, and also changed a few lines of code to reference the
live variables instead of test data.

Finally, I moved the "return YES" line of application:
didFinishLaunchingWithOptions: in UMAAppDelegate.m so that the rest of
the code would execute.