Working!
I've changed just a few things to get this working.

1) The feed interface was creating a new instance of the twitter
controller, so I changed it to get a reference to the singleton
instance instead. That also fixed the issue of the location service not
returning real values.

2) I've moved a couple of variable declaration to the headers, and
moved the instantiation of the location service in the twitter
controller to the init method. That way, the twitter controller keeps a
private instance of the location service running.